Add plug-in detection

+++ smartfancontrol/features/lapmode.py
@@ -1,6 +1,8 @@
 import tensorflow as tf
 
 
+# At least on my P14s gen 2 this sensor is not very accurate
+
 def read_lapmode():
     with open("/sys/module/thinkpad_acpi/drivers/platform:thinkpad_acpi/thinkpad_acpi/dytc_lapmode", "r", 1) as file:
         try:
@@ -10,4 +12,4 @@ def read_lapmode():
 
 
 def extract_lapmode_tensor(d: dict) -> dict:
-    return {k: tf.convert_to_tensor([v], dtype=tf.float32) for k, v in d.items()}
+    return {k: tf.convert_to_tensor([v], dtype=tf.int8) for k, v in d.items()}
